+++ src/test/java/com/tsp/new_tsp_front/api/portfolio/controller/FrontPortFolioJpaApiControllerTest.java
@@ -42,6 +42,7 @@ class FrontPortFolioJpaApiControllerTest {
     public void setup() {
         this.mockMvc = webAppContextSetup(wac)
                 .addFilters(new CharacterEncodingFilter("UTF-8", true))  // 필터 추가
+                .alwaysExpect(status().isOk())
                 .alwaysDo(print())
                 .build();
     }
