+++ src/test/java/com/tsp/new_tsp_front/api/support/controller/FrontSupportJpaApiControllerTest.java
@@ -44,6 +44,7 @@ class FrontSupportJpaApiControllerTest {
 	public void setup() {
 		this.mockMvc = webAppContextSetup(wac)
 				.addFilters(new CharacterEncodingFilter("UTF-8", true))  // 필터 추가
+				.alwaysExpect(status().isOk())
 				.alwaysDo(print())
 				.build();
 	}
